requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
constants==0.6.0
dj-database-url==1.0.0
Django==3.2.5
django-admin-volt==0.0.5
django-appconf==1.0.5
django-compressor==4.0
django-heroku==0.3.1
django-sass-processor==1.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.11.3
libsass==0.21.0
mccabe==0.7.0
Pillow==9.0.1
psycopg2==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2021.3
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
whitenoise==6.2.0
wincertstore==0.2
zipp==3.7.0
